pyproject.toml
[project]
name = "code-obfuscation-research"
version = "0.1.0"
description = "Measure utility drop from code perturbation for privacy protection"
requires-python = ">=3.12"
dependencies = [
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "langchain-core>=0.3",
    "langchain-community>=0.3",
    "langchain-openai>=0.3",
    "deepeval>=2.1",
    "datasets>=3.0",
    "libcst>=1.4",
    "pydantic>=2.7",
    "boilerplate-tools",
    "tqdm>=4.66",
    "rope>=1.14.0",
    "mini-swe-agent>=2.2.8",
    "swebench>=4.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-timeout>=2.3",
    "ruff>=0.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/code_obfuscation_research", "swebench_task"]

[tool.pytest.ini_options]
testpaths = ["tests", "swebench_task/tests"]
asyncio_mode = "auto"
timeout = 60

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src", "scripts", "tests", "swebench_task"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
